move some logic into market context
